refactor: Move commonalities Part 4
- XXX_to_hashmap moved
- file_rename moved (could be moved to a generic utilities crate)
