+++ utilities/CHANGELOG.md
@@ -20,6 +20,12 @@ All notable changes to this project will be documented in this file.
 - Changelog update
 - Changelog
 
+### Refactor
+
+- Move commonalities to shared package
+- Commonalities to shared package Part 2
+- Separate out commonalities Part 3
+
 ## [0.3.6] - 2022-01-03
 
 ### Chore
