+++ utilities/Cargo.toml
@@ -17,6 +17,7 @@ serde = { version = "1.0.133", features = ["derive"] }
 serde_json = "1.0.75"
 log = "0.4.14"
 csv = "1.1.6"
+convert_case = "0.5.0"
 
 # Fitness
 fitparser = "0.4.3"
