+++ utilities/src/fit/to_hashmap.rs
@@ -1,42 +1,77 @@
-use std::{collections::HashMap, error::Error, fs::File, io::BufReader};
+use crate::FITSession;
+// use utilities::{self};
 
 use chrono::{Datelike, Timelike};
-use gpx::Gpx;
+use convert_case::{Case, Casing};
+use fitparser::profile::field_types::MesgNum;
+use std::{collections::HashMap, error::Error, fs::File};
 
-use utilities::GPXMetadata;
-
-/// Parses a GPX file and returns the relevant metadata
-pub fn process_gpx(filename: &str) -> Result<HashMap<String, String>, Box<dyn Error>> {
-    let gpx: Gpx = gpx::read(BufReader::new(File::open(&filename)?))?;
-    log::debug!("process_gpx::gpx = {:?}", gpx);
-    let gpxmeta = GPXMetadata::from_header(&gpx, filename)?;
-    log::debug!("process_gpx::gpxmeta = {:?}", gpxmeta);
-
+/// Process a .FIT file and return the results
+pub fn fit_to_hashmap(filename: &str) -> Result<HashMap<String, String>, Box<dyn Error>> {
     let mut values = HashMap::<String, String>::new();
-    let mf = gpxmeta.creator.unwrap_or_else(|| "unknown".to_string());
-    values.insert("%manufacturer".to_string(), mf.clone());
-    values.insert("%mf".to_string(), mf.clone());
-    values.insert("%product".to_string(), mf.clone());
-    values.insert("%pr".to_string(), mf);
 
-    let sn = gpxmeta
-        .description
+    // open the file and deserialize it - return error if unable.
+    let mut fp = File::open(filename)?;
+    let file = fitparser::from_reader(&mut fp)?;
+
+    // Create a bunch of placeholder variables.
+    let mut my_session = FITSession::with_filename(filename)?;
+    let mut num_sessions = 0;
+
+    // This is where the actual parsing happens
+    for data in file {
+        // for each FitDataRecord
+        match data.kind() {
+            // Figure out what kind it is and parse accordingly
+            MesgNum::FileId => {
+                // File header
+                my_session.parse_header(data.fields())?;
+            }
+            MesgNum::Session => {
+                my_session.parse_session(data.fields())?;
+                num_sessions += 1;
+                my_session.num_sessions = Some(num_sessions);
+            }
+            _ => (),
+        } // match
+    } // for data
+
+    // Push the data into the HashMap for later use.
+    let mf = my_session
+        .manufacturer
         .unwrap_or_else(|| "unknown".to_string())
-        .replace("GPX File Created by ", "")
-        .trim()
-        .to_string();
+        .to_case(Case::Title);
+    values.insert("%manufacturer".to_string(), mf.clone());
+    values.insert("%mf".to_string(), mf);
+
+    let pr = my_session
+        .product
+        .unwrap_or_else(|| "unknown".to_string())
+        .to_case(Case::Title);
+    values.insert("%product".to_string(), pr.clone());
+    values.insert("%pr".to_string(), pr);
+
+    let sn = my_session
+        .serial_number
+        .unwrap_or_else(|| "unknown".to_string());
     values.insert("%serial_number".to_string(), sn.clone());
     values.insert("%sn".to_string(), sn);
 
-    let ac = gpxmeta.activity.unwrap_or_else(|| "unknown".to_string());
+    let ac = my_session
+        .activity_type
+        .unwrap_or_else(|| "unknown".to_string())
+        .to_case(Case::Title);
     values.insert("%activity".to_string(), ac.clone());
     values.insert("%ac".to_string(), ac);
 
-    let ad = "unknown".to_string();
+    let ad = my_session
+        .activity_detailed
+        .unwrap_or_else(|| "unknown".to_string())
+        .to_case(Case::Title);
     values.insert("%activity_detailed".to_string(), ad.clone());
     values.insert("%ad".to_string(), ad);
 
-    if let Some(tc) = gpxmeta.time {
+    if let Some(tc) = my_session.time_created {
         values.insert("%year".to_string(), format!("{:04}", tc.0.year()));
         values.insert("%yr".to_string(), format!("{:04}", tc.0.year()));
         values.insert("%month".to_string(), format!("{:02}", tc.0.month()));
@@ -90,7 +125,7 @@ pub fn process_gpx(filename: &str) -> Result<HashMap<String, String>, Box<dyn Er
         values.insert("%wd".to_string(), "00".to_string());
     }
 
-    if let Some(dur) = gpxmeta.duration {
+    if let Some(dur) = my_session.duration {
         values.insert("%duration".to_string(), dur.to_string());
         values.insert("%du".to_string(), dur.to_string());
     } else {
@@ -98,5 +133,24 @@ pub fn process_gpx(filename: &str) -> Result<HashMap<String, String>, Box<dyn Er
         values.insert("%du".to_string(), "0".to_string());
     }
 
+    // return safely
     Ok(values)
+}
+
+#[cfg(test)]
+///
+mod tests {
+    use super::*;
+    use assay::assay;
+
+    #[assay(include = ["/Users/evensolberg/CloudStation/Source/Rust/fitutils/data/test.fit"])]
+    ///
+    fn test_process_fit() {
+        let filename = "/Users/evensolberg/CloudStation/Source/Rust/fitutils/data/test.fit";
+        let fm = fit_to_hashmap(filename)?;
+
+        // File contents only get printed if run with cargo test -- --nocapture
+        println!("tm = {:?}", fm);
+        println!("tm.len() = {}", fm.len());
+    }
 }
